Add telepresence2 formula
This is so we can start brew installing telepresence2.  In the
telepresence.io cirlce ci, I will add the ability for it to update this
file.

Once we do a cutover on the docs on telepresence.io to have them point
at the newer docs, I will make another PR to this repo to change the
name so this formula is simply called telepresence.